fix: fix set docs by slice
